--- runtime/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "arcs-runtime",
-  "version": "0.0.0",
-  "private": "true",
-  "scripts": {
-    "test": "gulp"
-  },
-  "devDependencies": {
-    "chai": "^3.5.0",
-    "gulp": "^3.9.1",
-    "gulp-mocha": "^4.3.1",
-    "gulp-peg": "^0.2.0",
-    "gulp-util": "^3.0.8",
-    "mocha": "^3.2.0",
-    "node-fetch": "^1.7.2",
-    "pegjs": "^0.10.0",
-    "watchify": "^3.9.0",
-    "webpack": "^2.6.1",
-    "../tracelib/trace.js": "../tracelib"
-  },
-  "dependencies": {
-    "@std/esm": "^0.16.0",
-    "module-alias": "^2.0.2"
-  }
-}
