--- tracelib/package.json
@@ -1,10 +1,0 @@
-{
- "name": "tracelib",
- "version": "0.0.0",
- "private": "true",
- "main": "trace.js",
- "dependencies": {
-  "minimist": "*",
-  "mkdirp": "*"
- }
-}
